fixup! x4xx: x4xx commits go here
X410: Disable GPS sensors when GPS is disabled

UHD (mpmd) does not know about sensors that get dynamically
enabled/disabled. For that reason, we enable all sensors all the time.

However, the GPS sensors try and talk to the GPS, which can fail if GPS
is disabled. We thus check for GPS-enabled, and return dummy sensor
values if it's not.