fixup! x4xx: x4xx commits go here
X410: MB CPLD: Remove PRC-enabled warnings

The MB CPLD currently throws a warning if the PRC is already enabled
when trying to enable. However, there's no need for this. The only
important check is that the PRC is actually enabled (or disabled) after
the request.
The contract for the API call is clarified, and TODOs are removed.