fixup! x4xx: x4xx commits go here
X410/ZBX: Improve error message when choosing an invalid LO name,
including 'all'.

multi_usrp::get_rx_lo_source() and friends defaults to an LO name 'all'
(or ALL_LOS). This value is not really applicable to ZBX, and as such,
it's an unfortunate default value. We improve the error message.